chore: Extract `NeighborHoodIndices`
